% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_breakageparameter.R
\name{calc_breakageparameter}
\alias{calc_breakageparameter}
\title{Calculate current breakage parameter `fb` based on tensile stress}
\usage{
calc_breakageparameter(tr, tru, kappat = NULL, fb0 = NULL)
}
\arguments{
\item{tr}{current tensile stress in root (assuming unbroken) (array, size m)}

\item{tru}{root tensile strength (array, size m)}

\item{fb0}{Current value of breakage parameter that may not be exceeded
(array or scalar)}

\item{kappa}{Weibull shape parameter (array or scalar)}
}
\value{
array with new values of breakage parameter `fb`
}
\description{
Calculate the breakage parameter `fb`, i.e. the fraction of roots still
unbroken given the current conditions in the root. This fraction is 
calculated using the current tensile stress in the root. 
Weibull failure can be incorporated if a `kappa` parameter, a Weibull
shape parameter, is specified in the input. A current value of 
breakage parameter (`fb0`) that cannot be exceeded (to ensure roots
cannot 'unbreak' again) may be specified.
}
\examples{
calc_breakageparameter(c(1,2,3), 2.5)
calc_breakageparameter(c(1,2,3), 2.5, kappat = 4)
calc_breakageparameter(c(1,2,3), 2.5, kappat = 4, fb0 = c(0.6, 0.6, 0.6))
}
