% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cosalpha.R
\name{calc_cosalpha}
\alias{calc_cosalpha}
\title{Calculate cosine of azimuth of displaced orienation}
\usage{
calc_cosalpha(alpha0, beta0, us, h)
}
\arguments{
\item{alpha0}{azimuth angle of the initial root orientation}

\item{beta0}{elevation angle of the initial root orientation}

\item{us}{soil shear displacement}

\item{h}{shear zone thickness}
}
\value{
cosine of the displaced azimuth
}
\description{
Function that calculated the cosine of the azimuth angle describing
the displaced root orientation
}
\examples{
calc_cosalpha(pi/4, pi/4, 10, 30)
}
