% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cosbeta.R
\name{calc_cosbeta}
\alias{calc_cosbeta}
\title{Calculate cosine of elevation of displaced orienation}
\usage{
calc_cosbeta(alpha0, beta0, us, h)
}
\arguments{
\item{alpha0}{azimuth angle of the initial root orientation}

\item{beta0}{elevation angle of the initial root orientation}

\item{us}{shear displacement}

\item{h}{shear zone thickness}
}
\value{
cosine of the displaced elevation
}
\description{
Function that calculated the cosine of the elevation angle describing
the displaced root orientation
}
\examples{
calc_cosbeta(pi/4, pi/4, 10, 30)
}
