% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rootreinforcement.R
\name{calc_rootreinforcement}
\alias{calc_rootreinforcement}
\title{Calculate root reinforcement from known root stresses}
\usage{
calc_rootreinforcement(phi, tr, phir, cosalpha, sinbeta, cosbeta)
}
\arguments{
\item{phi}{soil angle of internal friction}

\item{tr}{tensile stress in root (including breakage effects)}

\item{phir}{root area ratio per root}

\item{cosalpha}{cosine of displaced root azimuth in shear zone}

\item{sinbeta}{sine of displaced root elevation in shear zone}

\item{cosbeta}{cosine of displaced root elevation in shear zone}
}
\value{
root reinforcement for every root in input
}
\description{
Calculates the current amount of root reinforcement for each root in
the input.
}
\examples{
calc_rootreinforcement(30*(2*pi/180), 10, 0.01, 1, sin(pi/5), cos(pi/6))
}
