% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sinbeta.R
\name{calc_sinbeta}
\alias{calc_sinbeta}
\title{Calculate sine of elevation of displaced orienation}
\usage{
calc_sinbeta(alpha0, beta0, us, h)
}
\arguments{
\item{alpha0}{azimuth angle of the initial root orientation}

\item{beta0}{elevation angle of the initial root orientation}

\item{us}{shear displacement}

\item{h}{shear zone thickness}
}
\value{
sine of the displaced elevation
}
\description{
Function that calculated the sine of the elevation angle describing
the displaced root orientation
}
\examples{
calc_sinbeta(pi/4, pi/4, 10, 30)
}
