% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cappedsphere_vertices.R
\name{cappedsphere_vertices}
\alias{cappedsphere_vertices}
\title{Get coordinates and vertices of a capped sphere}
\usage{
cappedsphere_vertices(
  alpha0 = c(-pi, pi),
  beta0 = c(0, pi),
  nalpha = 24,
  nbeta = 12,
  alphaoffset = 0,
  betaoffset = 0
)
}
\arguments{
\item{alpha0}{array with upper and lower azimuth limits}

\item{beta0}{array with upper and lower elevation limits}

\item{nalpha}{number of points to discretise alpha}

\item{nbeta}{number of points to discretise beta}

\item{alphaoffset}{azimuth offset of the apex of the capped sphere}

\item{betaoffset}{elevation offset of the apex of the capped sphere}
}
\value{
list with two dataframes, one for coordinates (fields
  `x`, `y` and `z`), and one with vertices (fields `i`, `j`,
  `k`, referring to the row of associated coordiantes)
}
\description{
Function to generate dataframes containing the coordinates on a unit
sphere and the vertices required to draw a unit semi-sphere
}
\examples{
cappedsphere_vertices()
cappedsphere_vertices(beta0 = c(0, pi/8))
}
