% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_allorientationsproperties.R
\name{create_allorientationsproperties}
\alias{create_allorientationsproperties}
\title{Create dataframe with all root properties and orientations}
\usage{
create_allorientationsproperties(dr, do)
}
\arguments{
\item{dr}{dataframe with root properties}

\item{do}{dataframe with orientation properties}
}
\value{
dataframe with all combinations of root properties and orientations
  required in DRAM
}
\description{
Function that creates dataframe with all root properties and orientations,
combining input already defined input for properties and orientations
}
\examples{
dr <- data.frame(
  phir = c(1, 2, 1),
  tru = c(2, 3, 4),
  dr = c(1, 2, 3)
)
do <- data.frame(
  alpha0 = c(0, 0, 1),
  beta0 = c(0, 1, 1),
  phir = c(1, 1.5, 1.5)
)
create_allorientationsproperties(dr, do)
}
