% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rootproperties.R
\name{create_rootproperties}
\alias{create_rootproperties}
\title{Create all root properties}
\usage{
create_rootproperties(
  drmin,
  drmax,
  nc,
  phirt,
  betaphi,
  Lr0,
  betaL,
  tru0,
  betat,
  trytru,
  epsru0,
  betaeps,
  epsryepsru,
  kappat,
  dr0 = 1,
  du = NULL
)
}
\arguments{
\item{drmin}{minimum root diamete (scalar)}

\item{drmax}{maximum root diameter (scalar)}

\item{nc}{number of root classes to use (integer scalar)}

\item{phirt}{total root area ratio}

\item{betaphi}{root area ratio power law coefficient (scalar)}

\item{Lr0}{root length of root with reference diameter (scalar)}

\item{betaL}{power law coefficient for root length (scalar)}

\item{tru0}{root tensile strength of root with reference diameter (scalar)}

\item{betat}{power law coefficient for root tensile strength (scalar)}

\item{trytru}{ratio of yield and tensile strength (scalar)}

\item{epsru0}{tensile strain to peak for root with reference diameter
(scalar)}

\item{betaeps}{power law coefficient for root tensile strain to
peak (scalar)}

\item{epsryepsru}{ratio of yield and tensile strain (scalar)}

\item{kappat}{Weibull shape parameter for tensile strength (scalar)}

\item{dr0}{reference diameter (scalar)}

\item{du}{dataframe with unit conversions}
}
\value{
dataframe with root diameter classes and root properties
}
\description{
Function that creates dataframe with root properties based on input.
}
\examples{
create_rootproperties(2,5, 10, 0.01,-0.2, 1,0.5, 10e6,-0.5,0.5, 0.2,0,0.1, 5)
}
