% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasave_inputparameters.R
\name{datasave_inputparameters}
\alias{datasave_inputparameters}
\title{Generate dataframe with all input parameters, values and units}
\usage{
datasave_inputparameters(input, norientation_used = NULL, du = NULL)
}
\arguments{
\item{input}{structure with all Shiny interactive input}

\item{norientation_used}{actual number of discrete root orientations used}

\item{du}{dataframe with unit conventions}
}
\value{
dataframe with all DRAM input properties in unit system defined
  by user
}
\description{
Function that takes the current input into the Shiny application and
generates a dataframe with parameter names, values and units.
}
