% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasave_outputparameters.R
\name{datasave_outputparameters}
\alias{datasave_outputparameters}
\title{Generate dataframe with all (summary) output parameters, values and units}
\usage{
datasave_outputparameters(dsum, du = NULL)
}
\arguments{
\item{dsum}{dataframe with output (in SI units)}

\item{du}{dataframe with unit conventions}
}
\value{
dataframe with all properties
}
\description{
Function that takes the (summary) output from a DRAM analysis and creates
a dataframe
}
