% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destabilising_shearstress.R
\name{destabilising_shearstress}
\alias{destabilising_shearstress}
\title{Calculate shear stress on soil outside shear zone}
\usage{
destabilising_shearstress(tr, phir, phi, cosalpha, sinbeta, cosbeta)
}
\arguments{
\item{tr}{soil (apparent) cohesion (array)}

\item{phir}{root area ratio (array)}

\item{phi}{soil angle of internal friction (scalar)}

\item{cosalpha}{cosine of azimuth angle of displaced root in shear zone}

\item{sinbeta}{sine of elevation angle of displaced root in shear zone}

\item{cosbeta}{cosine of elevation angle of displaced root in shear zone}
}
\value{
shear stress applied by roots on soil just outside the shear zone
}
\description{
Calculate the shear stress roots apply on the soil just outside the 
shear zone. This destabilising stress may result in an increase in 
the shear zone thickness if this stress exceeds the shear stress in
the soil
}
\examples{
destabilising_shearstress(10, 0.01, 30*(2*pi/360), cos(0.1), sin(0.2), cos(0.2))
}
