% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dram_runanalysis.R
\name{dram_runanalysis}
\alias{dram_runanalysis}
\title{DRAM calculation function}
\usage{
dram_runanalysis(da, ds, updateProgress = NULL)
}
\arguments{
\item{da}{dataframe with root properties and orientations. This should
contain fields for:
the total root area ratio (`phir`),
initial orientation azimuth (`alpha0`) [in radians],
initial orientation elevation (`beta0`) [in radians],
root circumference (`Cr`),
root cross-sectional area (`Ar`),
root length (`Lr`),
root yield strength (`try`),
root tensile strength (`tru`),
root elastic stiffness (`Ere`),
root elasto-plastic stiffness (`Erp`),
Weibull breakage parameter (`kappa`)}

\item{ds}{dataframe with soil properties. This should contain
fields for:
soil cohesion (`c`),
soil angle of internal friction (`phi`) [rad],
normal effective soil stress on shear plane (`sign`),
root-soil interface friction (`taui`),
initial shear zone thickness (`h0`),
maximum shear zone thickness (`hmax`),
maximum shear displacement (`usmax`),
number of steps  (`nstep`)}

\item{updateProgress}{shiny object to update calculation progress in
shiny UI}
}
\value{
List with two dataframes:
  a) `sum`, which contains summary data for each step.
  This contains fields:
  step identifier (`stepID`),
  shear displacement (`u`),
  shear zone thickness (`h`),
  root reinforcement (`cr`);
  b) `all`, which contains summary data for each step and root.
  This contains fields for:
  step identifier (`stepID`),
  root identified (`rootID`),
  root tensile strain assuming no root breakage (`tr`),
  breakage parameter/current fraction of roots intact (`fb`),
  root reinforcement per root (`cr`),
  a flag indicating the type of root behaviour (`flag`):
  `flag=0` = root not in tension,
  `flag=1` = anchored elastic behaviour,
  `flag=2` = anchored elastoplastic behaviour
  `flag=3` = slipping elastic behaviour,root not in tension,
  `flag=4` = slipping elastoplastic behaviour.
}
\description{
Function that takes all root properties/orientations and calculates
root reinforcements at every displacement step
}
\examples{
da <- data.frame(rootID=1, phir=0.01, alpha0=0, beta0=0, Cr=pi, Ar=pi/4, Lr=100, try=10e6, tru=20e6, Ere=100e6, Erp=50e6, kappat=5)
ds <- data.frame(c=1e3, phi=0.5, sign=10e3, taui=1e3, h0=1, hmax=50, usmax=50, nstep=100)
dram_runanalysis(da, ds)
}
