% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_grid_df.R
\name{expand_grid_df}
\alias{expand_grid_df}
\title{Expand.grid for dataframes}
\usage{
expand_grid_df(...)
}
\arguments{
\item{...}{dataframes}
}
\value{
dataframe all combinations of rows, merged
}
\description{
Equivalent of base-R `expand.grid` for dataframes. Each combination of 
rows in the dataframes is outputted
}
\examples{
d1 <- data.frame(a = c(1,2,3), b = c(2,3,4))
d2 <- data.frame(c = c(5,6), d = c(7,8))
expand_grid_df(d1, d2)
}
