% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientations_initial.R
\name{orientations_initial}
\alias{orientations_initial}
\title{Create discrete initial root orientations}
\usage{
orientations_initial(
  ndimension,
  norientation,
  phirt,
  beta0max,
  alphaoffset = 0,
  betaoffset = 0,
  input_volume = FALSE
)
}
\arguments{
\item{ndimension}{The number of dimensions to consider. `1` for 1D root
orientations (single orientation), `2` for 2-D orienations (orientation
all located on the same arc) or `3` for 3-D orientations (orientation
distributed on a spherical cap).}

\item{norientation}{The number of requested discrete root orienations.
if `ndimension==3` (3-D root orientation), the number of returned
orientations may be larger due to the discretisation method chosen.
See documentation for `initial_orientations_sphericalcap` for more
details}

\item{phirt}{total root area ratio on the shear plane (if `input_volume`)
is set to `TRUE`, this input is interpreted as the root volume ratio
instead. Input in terms of a fraction}

\item{beta0max}{the maximum (limiting) elevation angle for all root
orientations}

\item{alphaoffset}{azimuth of the (rotated) coordinate system used for initial
definition of orientations.}

\item{betaoffset}{elevation of the (rotated) coordinate system used for initial
definition of orientations.}

\item{input_volume}{if `TRUE`, the total root input `phirt` is interpreted
as the total root volume ratio in the soil volume.
If `FALSE` (default), the input is assumed to be the root area ratio
on the shear plane.}
}
\value{
dataframe with the initial azimuth (`alpha0`), the initial
  elevation (`beta0`), the relative 'weight' of each orientation, i.e.
  the fraction of root volume associated with each orientation
  (`weight`), and the root area ratio assigned to each orientation
  (`phir`)
}
\description{
Wrapper function to create a series of discrete, initial root orientations.
This function assumed a uniform distribution of roots in 1-D, (single
point), 2-D (orientations on the same arc) or 3-D (orientations on a
spherical grid).

First, the range of orientations is discretised into a number of discrete
orientations. Each orientation is assigned a weight, indicating the
fraction of all roots that is represented by that orientation.

Secondly, the root area ratio is calculated for each orientation, since
roots that are less perpendicular to the shear plane contribute less
towards the root area ratio.
}
\examples{
orientations_initial(1, 10, 0.01, pi/4)
orientations_initial(2, 10, 0.01, pi/4)
orientations_initial(3, 10, 0.01, pi/4)
orientations_initial(3, 10, 0.01, pi/4, input_volume = TRUE)
orientations_initial(3, 10, 0.01, pi/4, alphaoffset = 0.1, betaoffset = 0.1)
}
