% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientations_initial_arc.R
\name{orientations_initial_arc}
\alias{orientations_initial_arc}
\title{Create initial root orientations on a 2D range}
\usage{
orientations_initial_arc(beta0max, n)
}
\arguments{
\item{beta0max}{the maximum elevation angle describing the spherical cap}

\item{n}{number of discrete orientation requested}
}
\value{
dataframe with the initial azimuth (`alpha0`), the initial
  elevation (`beta0`) and the relative weight that should be assigned to
  each orientation (`weight`).
}
\description{
Determines discrete number of root orientations that approximate a
continuous distribution of initial root orienations in a 2D-plane.

The azimuth is assumed zero, and the elevations uniformly distributed
along an elevation range spanning from `-beta0max` to `beta0max`.
}
\examples{
orientations_initial_arc(pi/4, 6)
}
