% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientations_initial_sphericalcap.R
\name{orientations_initial_sphericalcap}
\alias{orientations_initial_sphericalcap}
\title{Create initial root orientations on a 3D-spherical cap}
\usage{
orientations_initial_sphericalcap(beta0max, n, band_offset = 0.5)
}
\arguments{
\item{beta0max}{the maximum elevation angle describing the spherical cap}

\item{n}{number of discrete orientation requested}

\item{band_offset}{offset each row of cells by a certain fraction of the
azimuth width of the cell}
}
\value{
dataframe with the initial azimuth (`alpha0`), the initial
  elevation (`beta0`) and the relative weight that should be assigned to
  each orientation (`weight`).
}
\description{
Determines discrete number of root orientations that approximate a
continuous distribution of initial root orienations on a spherical
cap described by a maximum elevation angle `beta0max`.

Root orientations are distributed among a range of equal-width
rings along the spherical cap. The number of cells in each band
increases with 4 for every ring, similar to an rectangular grid
of cells.

In each cell, a representative points is located such that the
point is located in the centre of gravity of each cell.

The number of returned discrete orientation may be larger than the
number requested (`n`) because of the grid method.

Each discrete orientation is assigned a certain `weight` according to the
area of the cell it represents. The sum of weights for all orientations
is 1.
}
\examples{
orientations_initial_sphericalcap(pi/4, 15)
orientations_initial_sphericalcap(pi/4, 25)
}
