% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientations_initial_sphericalcap_discretise.R
\name{orientations_initial_sphericalcap_discretise}
\alias{orientations_initial_sphericalcap_discretise}
\title{Split spherical cap into a number of discrete cells}
\usage{
orientations_initial_sphericalcap_discretise(beta0max, n, band_offset = 0.5)
}
\arguments{
\item{beta0max}{the maximum elevation angle describing the spherical cap}

\item{n}{number of discrete orientation requested}

\item{band_offset}{offset each row of cells by a certain fraction of the
azimuth width of the cell}
}
\value{
dataframe with an entry for each cell. For each, the band it belongs
  to (`band`), the cell number on the band (`cell`), the total number of cells
  in the current band (`nband`), and the lower and upper azimuth
  (`alpha0_0`, `alpha0_1`) and elevation angles (`beta0_0`, `beta0_1`) are returned
}
\description{
Determines discrete number of root orientations that approximate a
continuous distribution of initial root orienations on a spherical
cap described by a maximum elevation angle `beta0max`.

Root orientations are distributed among a range of equal-width
rings along the spherical cap. The number of cells in each band
increases with 4 for every ring, similar to an rectangular grid
of cells.

The number of returned discrete orientation may be larger than the
number requested (`n`) because of the grid method.
}
\examples{
orientations_initial_sphericalcap_discretise(pi/4, 15)
orientations_initial_sphericalcap_discretise(pi/4, 25)
}
