% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_behaviourfractions.R
\name{plotly_behaviourfractions}
\alias{plotly_behaviourfractions}
\title{Create plotly object for root behaviour types}
\usage{
plotly_behaviourfractions(
  dsum,
  du = NULL,
  nsignif = 3,
  plot_percentage = FALSE
)
}
\arguments{
\item{dsum}{dataframe with results. Should contain fields for the soil
shear displacement (`us`) fields for the fraction of roots that behave
not in tension (`frac_notintension`),
anchored elastic (`frac_anchoredelastic`),
anchored elastoplastic (`frac_anchoredelastoplastic`),
slipping elastic (`frac_slipelastic`),
slipping elastoplastic (`frac_slipelastoplastic`) or
broken (`frac_broken`)}

\item{du}{dataframe with unit system used}

\item{nsignif}{number of significant units used in plotly object}

\item{plot_percentage}{if `TRUE`, show fractions in percentages`}
}
\value{
plotly object
}
\description{
Function that creates a plotly object showing how much of the root area
ratio behaves according to a certain type ("anchored, elastic", "not in
tension", "slip, elastoplastic" etc) for each soil shear displacement
}
