% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_diameterclasses.R
\name{plotly_diameterclasses}
\alias{plotly_diameterclasses}
\title{Create plotly object showing root diameter classes}
\usage{
plotly_diameterclasses(dr, du = NULL, nsignif = 3)
}
\arguments{
\item{dr}{dataframe with root classes. Should contains fields for
the representative root diameter in the class (`dr`), the
minimum (`drmin`) and the maximum root diameter (`drmax`) in each
class as well as the root area ratio in each class (`phir`)}

\item{du}{dataframe with unit system used}

\item{nsignif}{number of significant units used in plotly object}
}
\value{
plotly object
}
\description{
Function that creates a plotly object showing the discretised
root diameter classes used in the model
}
\examples{
dr <- data.frame(
  dr = c(1.5, 2.5, 3.5),
  drmin = c(1, 2, 3),
  drmax = c(2, 3, 4),
  phir = c(0.01, 0.02, 0.03)
)
plotly_diameterclasses(dr)
}
