% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_existingmodelsbarchart.R
\name{plotly_existingmodelsbarchart}
\alias{plotly_existingmodelsbarchart}
\title{Create plotly with peak root reinforcement predictions}
\usage{
plotly_existingmodelsbarchart(dcru, du = NULL, nsignif = 3, ylim = c(0, NA))
}
\arguments{
\item{dcru}{dataframe with predictiondroot classes. Should contains
fields for the model name (`model`) and the peak root-reinforcement
prediction (`cru`).}

\item{du}{dataframe with unit system used}

\item{nsignif}{number of significant units used in plotly object}

\item{ylim}{array with minimum and maximum y-axis limits}
}
\value{
plotly object
}
\description{
Function that creates a plotly object showing the peak root
reinforcement according to various existing models
}
\examples{
dcru <- data.frame(
  model = c('WWM', 'FBM0'),
  cru = c(20, 10)
)
plotly_existingmodelsbarchart(dcru)
}
