% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_orientations_2D.R
\name{plotly_orientations_2D}
\alias{plotly_orientations_2D}
\title{Create plotly with discrete root orientation (polar coordinates)}
\usage{
plotly_orientations_2D(
  do,
  ndimension,
  beta0max,
  alphaoffset = 0,
  betaoffset = 0,
  dgrid = NULL,
  du = NULL,
  nsignif = 3
)
}
\arguments{
\item{do}{dataframe with details of discrete root orientations. Should
contain fields for the azimuth (`alpha0`), the elevation (`beta0`) and
the fraction of roots with this orientation (`weight`)}

\item{ndimension}{The number of dimensions to consider. `1` for 1D root
orientations (single orientation), `2` for 2-D orienations (orientation
all located on the same arc) or `3` for 3-D orientations (orientation
distributed on a spherical cap).}

\item{beta0max}{the maximum (limiting) elevation angle for all root
orientations}

\item{alphaoffset}{azimuth of the (rotated) coordinate system used for initial
definition of orientations.}

\item{betaoffset}{elevation of the (rotated) coordinate system used for initial
definition of orientations.}

\item{dgrid}{optional dataframe for drawing the boundaries of each grid
cell. If provided, should contain the fields `alpha0` and `beta0`.
line segments should be separated by `NA` values to plot a large number
of line segments in one command.}

\item{du}{dataframe with unit conventions}

\item{nsignif}{number of significant digits to use on Plotly hoverlabel}
}
\value{
plotly object
}
\description{
Function that creates a plotly object graphically showing the assumed
discrete root orientations in the model, in 2D polar coordinates
(azimuth, elevation).
}
\examples{
do <- data.frame(
  alpha0 = c(0, 1, 2),
  beta0 = c(0.5, 0.5, 0.7),
  weight = c(0.1, 0.1, 0.1)
)
plotly_orientations_2D(do, 3, 1)
plotly_orientations_2D(do, 3, 1, alphaoffset = 0.2, betaoffset = 0.3)
}
