% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_reinforcement.R
\name{plotly_reinforcement}
\alias{plotly_reinforcement}
\title{Create plotly object for soil displacement vs root reinforcement}
\usage{
plotly_reinforcement(dsum, du = NULL, nsignif = 3)
}
\arguments{
\item{dsum}{dataframe with results. Should contain fields for the soil
shear displacement (`us`) and the root reinforcement (`cr`)}

\item{du}{dataframe with unit system used}

\item{nsignif}{number of significant units used in plotly object}
}
\value{
plotly object
}
\description{
Function that creates a plotly object showing the root reinforcement
as function of direct shear displacement
}
\examples{
dsum <- data.frame(
  us = seq(0, 10, l = 11),
  cr = sqrt(seq(0, 10, l = 11))
)
plotly_reinforcement(dsum)
}
