% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_rootlength.R
\name{plotly_rootlength}
\alias{plotly_rootlength}
\title{Create plotly object showing root diameter vs root length}
\usage{
plotly_rootlength(dr, du = NULL, nsignif = 3)
}
\arguments{
\item{dr}{dataframe with root classes. Should contains fields for
the representative root diameter in the class (`dr`) and the
root length (`Lr`).}

\item{du}{dataframe with unit system used}

\item{nsignif}{number of significant units used in plotly object}
}
\value{
plotly object
}
\description{
Function that creates a plotly object showing the root length
in roots as function of their diameter
}
\examples{
dr <- data.frame(
  dr = c(1.5, 2.5, 3.5),
  Lr = c(100, 200, 300),
)
plotly_rootlength(dr)
}
