% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_shearzonethickness.R
\name{plotly_shearzonethickness}
\alias{plotly_shearzonethickness}
\title{Create plotly object for shear zone thickness vs root reinforcement}
\usage{
plotly_shearzonethickness(dsum, du = NULL, nsignif = 3)
}
\arguments{
\item{dsum}{dataframe with results. Should contain fields for the soil
shear displacement (`us`) and the shear zone thickness (`h`)}

\item{du}{dataframe with unit system used}

\item{nsignif}{number of significant units used in plotly object}
}
\value{
plotly object
}
\description{
Function that creates a plotly object showing the shear zone thickness
as function of direct shear displacement
}
\examples{
dsum <- data.frame(
  us = seq(0, 10, l = 11),
  h = sqrt(seq(0, 10, l = 11))
)
plotly_shearzonethickness(dsum)
}
