% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_soilyieldcriterion.R
\name{plotly_soilyieldcriterion}
\alias{plotly_soilyieldcriterion}
\title{Create plotly object for soil Mohr-Coulomb yield criterion}
\usage{
plotly_soilyieldcriterion(ds, du = NULL, nsignif = 4)
}
\arguments{
\item{ds}{dataframe with soil properties}

\item{du}{dataframe with unit system used}

\item{nsignif}{number of significant units used in plotly object}
}
\value{
plotly object
}
\description{
Function that creates a plotly object showing the peak soil
shear stress as function of the normal stress, following the
Mohr-Coulomb framework
}
\examples{
ds <- data.frame(
  c = 2,
  phi = 30/180*pi,
  sign = 5
)
plotly_soilyieldcriterion(ds)
}
