% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_stressstrain.R
\name{plotly_stressstrain}
\alias{plotly_stressstrain}
\title{Create plotly object for normalised stress-strain curves}
\usage{
plotly_stressstrain(dr, du = NULL, nsignif = 4)
}
\arguments{
\item{dr}{dataframe with root classes. Should contains fields for
the yield parameters (`try`, `epsry`) and peak parameters
(`tru`, `epsru`)}

\item{du}{dataframe with unit system used}

\item{nsignif}{number of significant units used in plotly object}
}
\value{
plotly object
}
\description{
Function that creates a plotly object showing the mobilisation
of tensile stress in the root as function of root tensile strain.
Stresses and strains are normalised by the stress and strain to
failure.
}
\examples{
dr <- data.frame(
  epsry = 0.05,
  epsru = 0.20,
  try = 5e6,
  tru = 10e6,
  Ere = 5e6 / 0.05,
  Erp = (10e6 - 5e6) / (0.20 - 0.05),
  kappat = 10
)
plotly_stressstrain(dr)
}
