% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_tensilestrain.R
\name{plotly_tensilestrain}
\alias{plotly_tensilestrain}
\title{Create plotly object showing root tensile strain to failure}
\usage{
plotly_tensilestrain(dr, du = NULL, nsignif = 3)
}
\arguments{
\item{dr}{dataframe with root classes. Should contains fields for
the representative root diameter in the class (`dr`) and the
representative tensile strain (`epsru`).}

\item{du}{dataframe with unit system used}

\item{nsignif}{number of significant units used in plotly object}
}
\value{
plotly object
}
\description{
Function that creates a plotly object showing the tensile
strain to failure in roots as function of their diameter
}
\examples{
dr <- data.frame(
  dr = c(1.5, 2.5, 3.5),
  epsru = c(0.1, 0.12, 0.11)
)
plotly_tensilestrain(dr)
}
