% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_tensilestrength.R
\name{plotly_tensilestrength}
\alias{plotly_tensilestrength}
\title{Create plotly object showing root tensile strengths}
\usage{
plotly_tensilestrength(dr, du = NULL, nsignif = 3)
}
\arguments{
\item{dr}{dataframe with root classes. Should contains fields for
the representative root diameter in the class (`dr`) and the
representative tensile strength (`tru`).}

\item{du}{dataframe with unit system used}

\item{nsignif}{number of significant units used in plotly object}
}
\value{
plotly object
}
\description{
Function that creates a plotly object showing the tensile
strength of roots as function of their diameters
}
\examples{
dr <- data.frame(
  dr = c(1.5, 2.5, 3.5),
  tru = c(1, 2, 3),
)
plotly_tensilestrength(dr)
}
