% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar2cartesian.R
\name{polar2cartesian}
\alias{polar2cartesian}
\title{Converts spherical to cartesian coordinates}
\usage{
polar2cartesian(alpha = NULL, beta = NULL, data = NULL)
}
\arguments{
\item{alpha}{azimuth angles (numerical array)}

\item{beta}{elevation angles (numerical array)}

\item{data}{dataframe with fields `alpha` and `beta`, or `alpha0`
and `beta0`. This dataframe
is used if alpha and beta not specifically specified. Results for
cartesian coordinates are calculated and added to the dataframe}
}
\value{
a dataframe with cartesian coordinates. Contains fields for
  x (`x`), y (`y`) and z (`z`) positions on the unit circle. In `data`
  is specified, these fields are added to the input dataframe and that
  dataframe is returned
}
\description{
Transform spherical coordinates, defined by an azimuth and elevation
angle, to 3D Cartesian coordinates on a unit circle.
}
\examples{
polar2cartesian(alpha = c(0,1,2), beta = c(1,1,1))
polar2cartesian(data = data.frame(alpha = c(0,1,2), beta = c(1,1,1)))
}
