% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar2polar.R
\name{polar2polar}
\alias{polar2polar}
\title{Transform polar coordinates to different system of polar coordinates}
\usage{
polar2polar(do, alphaoffset = 0, betaoffset = 0)
}
\arguments{
\item{do}{a dataframe with orienations. Should contain fields for the
azimuth (`alpha0`) and the elevation (`beta0`).}

\item{alphaoffset}{The azimuth rotation required to obtain the intial coordinate
system in terms of the coordinate system to be transformed to}

\item{betaoffset}{The elevation rotation required to obtain the intial coordinate
system in terms of the coordinate system to be transformed to}
}
\value{
a dataframe witht the same fields as input `do`, but with updated
  values for the azimuth (`alpha0`) and elevation angle (`beta0`)
}
\description{
function to calculate initial root angles by transforming angles in a
rotated coordinate system back to a coordinate system of interest.

The orientation of the initial coordinate system is descrbined in terms
of the system of interest by an azimuth angle `alphaoffset` and an elevation
angle `betaoffset`.
}
\examples{
do <- data.frame(
  alpha0 = c(-1, 0, 1, 1),
  beta0 = c(1, 1, 1.5, 1.2),
  weight = c(0.25, 0.25, 0.25, 0.25)
)
polar2polar(do, alphaoffset = 0.1, betaoffset = 0.1)
}
