% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_roottypes.R
\name{postprocess_roottypes}
\alias{postprocess_roottypes}
\title{Postprocess DRAM results to get fractions of behaviour type}
\usage{
postprocess_roottypes(dsum, dall, da)
}
\arguments{
\item{dsum}{dataframe with DRAM output per step. Should contain fields the
displacement step identifier `stepID`.}

\item{dall}{dataframe with DRAM outpput per step and root. Should contain
fields for the displacement step identifier (`stepID`), root identifier
(`rootID`), breakage parameter (`fb`) and the flag indicating the
type of root behaviour (`flag`)}

\item{da}{dataframe with root properties and orientation. Should contain
fields for the root identfier (`rootID`) and the root area ratio per
root (`phir`)}
}
\value{
dataframe `dsum` with added fields for the fraction of the root
  area ratio belonging to broken roots (`fraction_broken`),
  roots not in tension (`fraction_notintension`),
  anchored elastic roots (`fraction_anchoredelastic`),
  anchored elastoplastic roots (`fraction_anchoredelastoplastic`),
  slipping elastic roots (`fraction_slipelastic`) and
  slipping elastoplastic roots (`fraction_slipelastoplastic`)
}
\description{
Function that analyses the DRAM results and determines for each displacement
step what fraction of the root area ratio corresponds with the various types
of root behaviour (not in tension, anchored elastic, slipping elastoplastic
etc.). The results are added to the `dsum` dataframe that is inputted.
}
\examples{
dsum <- data.frame(
  stepID = c(1, 2)
)
dall <- data.frame(
  stepID = c(1, 1, 2, 2),
  rootID = c(1, 2, 1, 2),
  fb = c(1, 1, 0.9, 0.8),
  flag = c(0, 0, 1, 3)
)
da <- data.frame(
  rootID = c(1, 2),
  phir = c(0.01, 0.02)
)
postprocess_roottypes(dsum, dall, da)
}
