% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_existingmodels_fbmcw.R
\name{shiny_existingmodels_fbmcw}
\alias{shiny_existingmodels_fbmcw}
\title{Calculate root reinforcement according to FBMc and FBMcw model}
\usage{
shiny_existingmodels_fbmcw(
  drmin,
  drmax,
  betaF,
  phirt,
  betaphi,
  tru0,
  betat,
  kappa,
  du = NULL,
  k = 1.2,
  dr0 = 1
)
}
\arguments{
\item{drmin, drmax}{minumum and maximum root diameter in bundle
(numeric array)}

\item{betaF}{Load sharing coefficient (numeric array)}

\item{phirt}{Total root area ratio (numeric array)}

\item{betaphi}{Power law coefficient for root diameter-root area ratio
fit (numeric array)}

\item{tru0}{tensile strength in reference root (numeric array)}

\item{betat}{Power law coefficient for root diameter-root tensile strength
fit (numeric array)}

\item{kappa}{Weibull shape parameter for the root survival function
(numeric array)}

\item{du}{dataframe with Shiny unit system}

\item{k}{WWM orientation factor (default k = 1.2)}

\item{dr0}{Reference diameter (numeric scalar, default 1.0)}
}
\value{
dataframe with fields for models (`model`) and peak root
  reinforcement predictions (`cru`)
}
\description{
Wrapper function to calculate the peak root reinforcement according to
the FBMc and FBMcw models.
}
\examples{
shiny_existingmodels_fbmcw(1, 5, 1.5, 0.01, -1, 10e6, -0.5, 2.5)
}
