% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_existingmodels_fbmw.R
\name{shiny_existingmodels_fbmw}
\alias{shiny_existingmodels_fbmw}
\title{Calculate root reinforcement according to FBM and FBMw model}
\usage{
shiny_existingmodels_fbmw(
  dr,
  phir,
  tru,
  tru0,
  betaF,
  kappa,
  du = NULL,
  k = 1.2,
  dr0 = 1
)
}
\arguments{
\item{dr}{discrete root diameters (numeric array)}

\item{phir}{root area ratio per diameter class (numeric array)}

\item{tru}{root tensile strength per diameter class (numeric array)}

\item{tru0}{tensile strength in reference root (numeric array)}

\item{betaF}{FBM load sharing coefficient}

\item{kappa}{Weibull shape parameter for the root survival function
(numeric array)}

\item{du}{dataframe with Shiny unit system}

\item{k}{WWM orientation factor (default k = 1.2)}

\item{dr0}{Reference diameter (numeric scalar, default 1.0)}
}
\value{
dataframe with fields for models (`model`) and peak root
  reinforcement predictions (`cru`)
}
\description{
Wrapper function to calculate the peak root reinforcement according to
the FBM and FBMw models.
}
\examples{
shiny_existingmodels_fbmw(c(1,2,3), rep(0.01,3), 10e6*c(1,2,3), 10e6, 1.5, 2.5)
}
