% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_existingmodels_literature.R
\name{shiny_existingmodels_literature}
\alias{shiny_existingmodels_literature}
\title{Calculate root reinforcement according to existing reinforcement models}
\usage{
shiny_existingmodels_literature(
  dr,
  phir,
  tru,
  tru0,
  betat,
  betaeps,
  betaL,
  kappa,
  du = NULL,
  k = 1.2,
  dr0 = 1
)
}
\arguments{
\item{dr}{discrete root diameters (numeric array)}

\item{phir}{root area ratio per diameter class (numeric array)}

\item{tru}{root tensile strength per diameter class (numeric array)}

\item{tru0}{tensile strength in reference root (numeric array)}

\item{betat}{Root strength power law coefficient}

\item{betaeps}{Root strain power law coefficient}

\item{betaL}{Root length power law coefficient}

\item{kappa}{Weibull shape parameter for the root survival function
(numeric array)}

\item{du}{dataframe with Shiny unit system}

\item{k}{WWM orientation factor (default k = 1.2)}

\item{dr0}{Reference diameter (numeric scalar, default 1.0)}
}
\value{
dataframe with fields for models (`model`) and peak root
  reinforcement predictions (`cru`)
}
\description{
Wrapper function to calculate the peak root reinforcement according to
various FBM models, the WWM model and the RBMw model
}
\examples{
shiny_existingmodels_literature(c(1,2,3), rep(0.01,3), 10e6*c(1,2,3), 10e6, -0.5, 0, 1, 2.5)
}
