% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_unitsystem.R
\name{shiny_unitsystem}
\alias{shiny_unitsystem}
\title{Generate unit system in Shiny based on user-input}
\usage{
shiny_unitsystem(
  dp,
  lengthunit_radio,
  rootstressunit_radio,
  soilstressunit_radio,
  rootstrainunit_radio,
  rootarearatiounit_radio,
  angleunit_radio
)
}
\arguments{
\item{lengthunit_radio}{Shiny input from radio button selection for
length unit}

\item{rootstressunit_radio}{Shiny input from radio button selection
for root stress/strength/stiffness unit}

\item{soilstressunit_radio}{Shiny input from radio button selection
for soil stress/strength/stiffness unit}

\item{rootstrainunit_radio}{Shiny input from radio button selection
for root strain unit}

\item{rootarearatiounit_radio}{Shiny input from radio button selection
for root area ratio unit}

\item{angleunit_radio}{Shiny input from radio button selection
for angles (degrees or radians)}
}
\value{
returns a dataframe with multiplication factor to get from
  input unit to SI unit (`factor`), the user-chosen unit (`unit_user`),
  the SI unit (`unit_SI`). The rownames of the dataframe refer to the
  parameter of interest
}
\description{
Defines a unit system and conversions for Shiny parameters
}
\examples{
dp <- data.frame(
  parameter = c('tru0', 'Lr0', 'kappat'),
  unit_type = c('rootstress', 'rootlength', 'dimensionless'),
  unit_si = c('Pa', 'm', '-')
)
row.names(dp) <- dp$parameter
shiny_unitsystem(dp, 1, 3, 1, 1, 1, 1)
}
