% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_shearstrength.R
\name{soil_shearstrength}
\alias{soil_shearstrength}
\title{Calculate shear strength of soil}
\usage{
soil_shearstrength(c, phi, sign)
}
\arguments{
\item{c}{soil (apparent) cohesion}

\item{phi}{soil angle of internal friction}

\item{sign}{normal stress acting on the shear plane}
}
\value{
soil shear strength
}
\description{
Calculate the shear strength of soil according to the Mohr-Coulomb
failure criterion
}
\examples{
soil_shearstrength(1, 30*(2*pi/360), 10)
}
