% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_polynomial_cubic.R
\name{solve_polynomial_cubic}
\alias{solve_polynomial_cubic}
\title{Returns the largest real root of a cubic function}
\usage{
solve_polynomial_cubic(a, b, c, d)
}
\arguments{
\item{a}{cubic coefficient (numeric array)}

\item{b}{quadratic coefficient (numeric array)}

\item{c}{linear coefficient (numeric array)}

\item{d}{constant coefficient (numeric array)}
}
\value{
positive, real solution to the cubic equation (numeric array)
}
\description{
This function returns the largest real root of the cubic function:
  `a*x^3 + b*x^2 + c*x^2 + d == 0`
with the following assumptions that follow from the mobilisation of 
root stresses:
  `a,b,c >= 0` and `d <= 0`
because of these assumptions, there always exists a real, positive root
}
\examples{
solve_polynomial_cubic(1,2,3,-4)
}
