% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_polynomial_quadratic.R
\name{solve_polynomial_quadratic}
\alias{solve_polynomial_quadratic}
\title{Returns the largest real root of a quadratic function}
\usage{
solve_polynomial_quadratic(a, b, c)
}
\arguments{
\item{a}{quadratic coefficient (numeric array)}

\item{b}{linear coefficient (numeric array)}

\item{c}{constant coefficient (numeric array)}
}
\value{
largest solution to the quadratic equation (numeric array)
}
\description{
This function returns the largest root of the cubic function:
  `a*x^2 + b*x^2 + c == 0`
with the following assumptions that follow from the mobilisation of 
root stresses:
  `a < 0`
because of these assumptions, there always exists a real, positive root
}
\examples{
solve_polynomial_quadratic(-2,6,8)
}
