% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensilestrain_unbroken.R
\name{tensilestrain_unbroken}
\alias{tensilestrain_unbroken}
\title{Tensile strain in roots based on stress}
\usage{
tensilestrain_unbroken(tr, try, Ere, Erp)
}
\arguments{
\item{tr}{root tensile stress (array)}

\item{try}{root yield stress (array or scalar)}

\item{Ere}{root elastic stiffness (array or scalar)}

\item{Erp}{root elasto-plastic stiffness (array or scalar)}
}
\value{
tensile strain (array)
}
\description{
Function to calculate the tensile strain in a root with a bilinear
stress-strain behaviour (elasto-plastic).
}
\examples{
tensilestrain_unbroken(c(1,2,3,4), 2.5, 20, 10)
}
