% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensilestress_anchored_elastic.R
\name{tensilestress_anchored_elastic}
\alias{tensilestress_anchored_elastic}
\title{Tensile stress in anchored, elastic root}
\usage{
tensilestress_anchored_elastic(Cr, Ar, Ere, taui, h, cosbeta, cosbeta0)
}
\arguments{
\item{Cr}{root circumference (array)}

\item{Ar}{root cross-sectional area (array)}

\item{Ere}{root elastic stiffness (array)}

\item{taui}{root-soil interface shear stress (array)}

\item{h}{thickness of the shear plane (array)}

\item{cosbeta}{cosine of the elevation angle of the current orientation
of the root within the shear zone}

\item{cosbeta0}{cosine of the elevation angle of the initial orientation
of the root within the shear zone}
}
\value{
tensile stress in middle of root under the stated assumptions
}
\description{
Function calculates the tensile stress in a root where is crosses the shear
zone in the soil. It is assumed that the root is anchored (i.e. the ends 
do not move) and that the middle of the root is behaving elastic.
The root is assumed to be straight, unbranched, have two freely supported
ends and has the middle of the root located in the shear zone.
}
\examples{
tensilestress_anchored_elastic(pi, pi/4, 100, 0.002, 30, 0.8, 0.90)
}
