% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensilestress_slip_elastoplastic.R
\name{tensilestress_slip_elastoplastic}
\alias{tensilestress_slip_elastoplastic}
\title{Tensile stress in slipping, elasto-plastic root}
\usage{
tensilestress_slip_elastoplastic(Cr, Ar, Lr, try, Ere, Erp, taui, h, cosbeta)
}
\arguments{
\item{Cr}{root circumference (array)}

\item{Ar}{root cross-sectional area (array)}

\item{Lr}{root length (array)}

\item{Ere}{root elastic stiffness (array)}

\item{Erp}{root plastic stiffness (array)}

\item{taui}{root-soil interface shear stress (array)}

\item{h}{thickness of the shear plane (array)}

\item{cosbeta}{cosine of the elevation angle of the current orientation
of the root within the shear zone}
}
\value{
tensile stress in middle of root under the stated assumptions
}
\description{
Function calculates the tensile stress in a root where is crosses the shear
zone in the soil. It is assumed that the root slips (i.e. the ends move)
and that the middle of the root is behaving elasto-plastic.
The root is assumed to be straight, unbranched, have two freely supported
ends and has the middle of the root located in the shear zone.
}
\examples{
tensilestress_slip_elastoplastic(pi, pi/4, 200, 10, 100, 80, 0.002, 30, 0.8)
}
