% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rootproperties.R
\name{create_rootproperties}
\alias{create_rootproperties}
\title{Add root properties to a dataframe}
\usage{
create_rootproperties(
  d,
  Lr0,
  betaL,
  tru0,
  betat,
  trytru,
  epsru0,
  betaeps,
  epsryepsru,
  kappat,
  dr0 = 1,
  du = NULL
)
}
\arguments{
\item{d}{dataframe with root classes. Should contain field (`dr`) for
the representative root diameter in the class. This diameter is already
converted to a SI unit system}

\item{Lr0}{root length of root with reference diameter (scalar)}

\item{betaL}{power law coefficient for root length (scalar)}

\item{tru0}{root tensile strength of root with reference diameter (scalar)}

\item{betat}{power law coefficient for root tensile strength (scalar)}

\item{trytru}{ratio of yield and tensile strength (scalar)}

\item{epsru0}{tensile strain to peak for root with reference diameter
(scalar)}

\item{betaeps}{power law coefficient for root tensile strain to
peak (scalar)}

\item{epsryepsru}{ratio of yield and tensile strain (scalar)}

\item{kappat}{Weibull shape parameter for tensile strength (scalar)}

\item{dr0}{reference diameter (scalar)}

\item{du}{dataframe with unit conversions}
}
\value{
dataframe with root diameter classes and root properties
}
\description{
Function that adds various root properties to a dataframe containing
root diameter classes
}
\examples{
d <- data.frame(dr = c(0.001, 0.002, 0.003, 0.004))
create_rootproperties(d, 1, 0.5, 10e6, -0.5, 0.5, 0.2, 0, 0.1, 5)
}
